Implement footnotes migration logic and improve rendering security. Introduced a custom hook for managing footnotes from block attributes and post meta, ensuring backward compatibility. Enhanced rendering by escaping attributes and content for improved security.
